+++ src/crontab/src/ConfigProvider.php
@@ -14,9 +14,8 @@ namespace Hyperf\Crontab;
 
 use Hyperf\Crontab\Listener\CrontabRegisterListener;
 use Hyperf\Crontab\Listener\OnPipeMessageListener;
-use Hyperf\Crontab\Strategy\CoroutineStrategy;
-use Hyperf\Crontab\Strategy\ProcessStrategy;
 use Hyperf\Crontab\Strategy\StrategyInterface;
+use Hyperf\Crontab\Strategy\WorkerStrategy;
 
 class ConfigProvider
 {
@@ -24,8 +23,7 @@ class ConfigProvider
     {
         return [
             'dependencies' => [
-                // StrategyInterface::class => CoroutineStrategy::class,
-                StrategyInterface::class => ProcessStrategy::class,
+                StrategyInterface::class => WorkerStrategy::class,
             ],
             'listeners' => [
                 CrontabRegisterListener::class,