+++ src/http-server/tests/CoreMiddlewareTest.php
@@ -12,12 +12,18 @@ declare(strict_types=1);
 
 namespace HyperfTest\HttpServer;
 
+use Hyperf\HttpServer\CoreMiddleware;
 use Hyperf\HttpServer\Router\DispatcherFactory;
+use Hyperf\Utils\Contracts\Arrayable;
+use Hyperf\Utils\Contracts\Jsonable;
 use HyperfTest\HttpServer\Stub\CoreMiddlewareStub;
 use HyperfTest\HttpServer\Stub\DemoController;
 use Mockery;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use ReflectionMethod;
 
 /**
  * @internal
@@ -33,6 +39,60 @@ class CoreMiddlewareTest extends TestCase
         $params = $middleware->parseParameters(DemoController::class, 'index', ['id' => $id]);
 
         $this->assertSame([$id, 'Hyperf', []], $params);
+    }
+
+    public function testTransferToResponse()
+    {
+        $middleware = new CoreMiddlewareStub($container = $this->getContainer(), 'http');
+        $reflectionMethod = new ReflectionMethod(CoreMiddleware::class, 'transferToResponse');
+        $reflectionMethod->setAccessible(true);
+        $request = Mockery::mock(ServerRequestInterface::class);
+        /** @var ResponseInterface $response */
+
+        // String
+        $response = $reflectionMethod->invoke($middleware, $body = 'foo', $request);
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+        $this->assertSame($body, $response->getBody()->getContents());
+        $this->assertSame('text/plain', $response->getHeaderLine('content-type'));
+
+        // Array
+        $response = $reflectionMethod->invoke($middleware, $body = ['foo' => 'bar'], $request);
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+        $this->assertSame(json_encode($body), $response->getBody()->getContents());
+        $this->assertSame('application/json', $response->getHeaderLine('content-type'));
+
+        // Arrayable
+        $response = $reflectionMethod->invoke($middleware, new class() implements Arrayable {
+            public function toArray(): array
+            {
+                return ['foo' => 'bar'];
+            }
+        }, $request);
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+        $this->assertSame(json_encode(['foo' => 'bar']), $response->getBody()->getContents());
+        $this->assertSame('application/json', $response->getHeaderLine('content-type'));
+
+        // Jsonable
+        $response = $reflectionMethod->invoke($middleware, new class() implements Jsonable {
+            public function __toString(): string
+            {
+                return json_encode(['foo' => 'bar'], JSON_UNESCAPED_UNICODE);
+            }
+        }, $request);
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+        $this->assertSame(json_encode(['foo' => 'bar']), $response->getBody()->getContents());
+        $this->assertSame('application/json', $response->getHeaderLine('content-type'));
+
+        // __toString
+        $response = $reflectionMethod->invoke($middleware, new class() {
+            public function __toString(): string
+            {
+                return 'This is a string';
+            }
+        }, $request);
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+        $this->assertSame('This is a string', $response->getBody()->getContents());
+        $this->assertSame('text/plain', $response->getHeaderLine('content-type'));
     }
 
     protected function getContainer()