+++ src/utils/src/Functions.php
@@ -127,7 +127,7 @@ if (! function_exists('data_get')) {
     /**
      * Get an item from an array or object using "dot" notation.
      *
-     * @param array|string $key
+     * @param array|int|string $key
      * @param null|mixed $default
      * @param mixed $target
      */
@@ -136,7 +136,8 @@ if (! function_exists('data_get')) {
         if (is_null($key)) {
             return $target;
         }
-        $key = is_array($key) ? $key : explode('.', $key);
+
+        $key = is_array($key) ? $key : explode('.', is_int($key) ? (string) $key : $key);
         while (! is_null($segment = array_shift($key))) {
             if ($segment === '*') {
                 if ($target instanceof Collection) {
@@ -413,7 +414,7 @@ if (! function_exists('run')) {
     /**
      * Run callable in non-coroutine environment, all hook functions by Swoole only available in the callable.
      */
-    function run(callable $callback, $flags = SWOOLE_HOOK_ALL): bool
+    function run(callable $callback, int $flags = SWOOLE_HOOK_ALL): bool
     {
         if (Coroutine::inCoroutine()) {
             throw new RuntimeException('Function \'run\' only execute in non-coroutine environment.');