+++ src/utils/tests/FunctionTest.php
@@ -47,6 +47,17 @@ class FunctionTest extends TestCase
         $this->assertSame('hyperf', $result);
     }
 
+    public function testDateGetIntegerKey()
+    {
+        $data = [1, 2, 3];
+        $result = data_get($data, 0);
+        $this->assertSame(1, $result);
+
+        $data = ['a' => [1, 2, 3], 4];
+        $result = data_get($data, 0);
+        $this->assertSame(4, $result);
+    }
+
     /**
      * @expectedException \HyperfTest\Utils\Exception\RetryException
      */